refactor: 이동 로직 코드 분리(service, controller)
