+++ src/main/resources/templates/chess.html
@@ -161,11 +161,11 @@
             <input class="btn btn-danger" type="submit" value="End"/>
           </form>
         </div>
-        <div class="col-md-4 button status-button">
-          <form action="/status" method="get">
-            <input class="btn btn-success" type="submit" value="Status"/>
-          </form>
-        </div>
+<!--        <div class="col-md-4 button status-button">-->
+<!--          <form action="/status" method="get">-->
+<!--            <input class="btn btn-success" type="submit" value="Status"/>-->
+<!--          </form>-->
+<!--        </div>-->
       </div>
       <div class="row menu-title">
         <div class="col-md-4 ">
