+++ src/main/resources/templates/index.html
@@ -13,7 +13,7 @@
 <div class="container">
   <div class="row">
     <div class="col">
-      <form action="/play" method="post">
+      <form action="/start" method="post">
         <input class="name-box" type="text" name="gameName"/>
         <input class="btn btn-default" type="submit" value="Start"/>
       </form>
